updating syntax/fixing composter string

+++ src/api/java/com/minecolonies/api/configuration/Configurations.java
@@ -202,13 +202,13 @@ public class Configurations
         public boolean rangerEnchants = true;
 
         @Config.Comment("Damage multiplier for Ranger Guards: Default:1.0")
-        public float rangerDamageMult = 1.0f;
+        public double rangerDamageMult = 1.0;
 
         @Config.Comment("Damage multiplier for Knight Guards: Default:1.0")
-        public float knightDamageMult = 1.0f;
+        public double knightDamageMult = 1.0;
 
         @Config.Comment("Health multiplier for all Guards: Default:1.0")
-        public float guardHealthMult = 1.0f;
+        public double guardHealthMult = 1.0;
 
         @Config.Comment("Amount of blocks the builder checks (to decrease lag by builder)")
         public int maxBlocksCheckedByBuilder = 1000;
